Migrate color prototype to common/color
