+++ game/domain/definitions/colors.lua
@@ -1,25 +1,5 @@
 
-local Color = require 'lux.prototype' :new { __type = 'color' }
-
-function Color.__mul(a, b)
-  if type(a) == 'table' and a.__type == 'color' and type(b) == 'table' and b.__type == 'color' then
-    return Color:new { a[1] * b[1] / 255,
-                      a[2] * b[2] / 255,
-                      a[3] * b[3] / 255,
-                      a[4] * b[4] / 255
-    }
-  elseif type(a) == 'table' and a.__type == 'color' and type(b) == 'number' then
-    return Color:new { a[1] * b, a[2] * b, a[3] * b, a[4] * b }
-  elseif type(b) == 'number' and type(b) == 'table' and b.__type == 'color' then
-    return Color:new { b[1] * a, b[2] * a, b[3] * a, b[4] * a }
-  end
-end
-
-function Color:__init()
-  for i = 1, 4 do
-    self[i] = self[i] or 0xff
-  end
-end
+local Color = require 'common.color'
 
 local COLORS = {}
